refactor: moved utils to separate files
