+++ plugin/src/generators/form/generator.ts
@@ -1,98 +1,9 @@
 import { addDependenciesToPackageJson, formatFiles, generateFiles, installPackagesTask, Tree } from '@nx/devkit';
 import * as path from 'path';
 import { FormGeneratorSchema } from './schema';
-import {
-  addNamedImport,
-  appendFileContent,
-  dynamicImport,
-  filterSource,
-  formatName,
-  getNxLibsPaths,
-  LibraryType,
-  searchAliasPath,
-  searchNxLibsPaths
-} from '../../shared/utils';
-import { existsSync } from 'fs';
-import { kebabCase } from 'lodash';
+import { dynamicImport, formatName, getNxLibsPaths, LibraryType, searchNxLibsPaths } from '../../shared/utils';
 import { dependencies } from '../../shared/dependencies';
-import {
-  ArrowFunction,
-  FunctionDeclaration,
-  FunctionExpression,
-  IndentationText,
-  Project,
-  QuoteKind,
-  SourceFile,
-  ts
-} from 'ts-morph';
-import SyntaxKind = ts.SyntaxKind;
-import { getPlaceOfUse } from './utils';
-
-async function getFormUtilsDirectory(): Promise<string> {
-  const { default: autocomplete } = await dynamicImport<typeof import('inquirer-autocomplete-standalone')>('inquirer-autocomplete-standalone');
-
-  const utilsLibsPaths = getNxLibsPaths([LibraryType.UTILS]);
-  const formUtilsLibsPaths = searchNxLibsPaths(utilsLibsPaths, 'utils/form/src', 'endsWith');
-
-  if (!formUtilsLibsPaths.length) {
-    throw new Error('Could not find a library with the form utilities.');
-  }
-
-  if (formUtilsLibsPaths.length > 1) {
-    formUtilsLibsPaths[0] = await autocomplete({
-      message: 'Select the path of the library with the form utilities: ',
-      source: (input) => filterSource(input, formUtilsLibsPaths)
-    });
-  }
-
-  return searchAliasPath(formUtilsLibsPaths[0]);
-}
-
-function updateIndex(formsPath: string, fileName: string, tree: Tree): void {
-  const formsIndexFilePath = `${formsPath}/index.ts`;
-  const newIndexContent = `export * from './${kebabCase(fileName)}';\n`;
-  if (!existsSync(formsIndexFilePath)) {
-    tree.write(formsIndexFilePath, newIndexContent);
-  } else {
-    appendFileContent(formsIndexFilePath, newIndexContent, tree);
-  }
-}
-
-function getFormUsageCode(formClassName: string): string {
-  return `const formSchema = new ${formClassName}();
-const form = useForm<${formClassName}>({
-  defaultValues: formSchema.formValues,
-  resolver: yupResolver<any>(${formClassName}.validationSchema)
-});\n\n`
-}
-
-async function addFormUsage(libPath: string, placeOfUseName: string, formClassName: string): Promise<void> {
-  const project = new Project({
-    manipulationSettings: {
-      indentationText: IndentationText.TwoSpaces,
-      quoteKind: QuoteKind.Single
-    }
-  });
-  const files = project.addSourceFilesAtPaths([`${libPath}/lib/**/*.tsx`, `${libPath}/lib/**/*.ts`]);
-  const file = files.find((file) => file.getFunction(placeOfUseName) || file.getVariableDeclaration(placeOfUseName));
-  if (!file) {
-    throw new Error('Could not find the place where the form should be used.');
-  }
-
-  const pathToForm = file.getFilePath().includes('components')
-    ? '../../forms'
-    : file.getFilePath().includes('hooks')
-      ? '../forms'
-      : './forms';
-  addNamedImport(formClassName, pathToForm, file);
-  addNamedImport('useForm', 'react-hook-form', file);
-  addNamedImport('yupResolver', '@hookform/resolvers/yup', file);
-
-  const placeOfUse = getPlaceOfUse(file, placeOfUseName);
-  placeOfUse.setBodyText(`${getFormUsageCode(formClassName)}${placeOfUse.getBodyText()}`);
-
-  project.saveSync();
-}
+import { addFormUsage, getFormUtilsDirectory, updateIndex } from './utils';
 
 export async function formGenerator(tree: Tree, options: FormGeneratorSchema) {
   // Get input data
