+++ plugin/src/generators/form/utils/index.ts
@@ -1,1 +1,3 @@
-export * from './get-place-of-use';
+export * from './get-form-utils-directory';
+export * from './update-index';
+export * from './add-form-usage';
